third_party/blink/web_tests/virtual/css-calc-infinity-and-nan-disabled/external/wpt/css/css-values/acos-asin-atan-atan2-serialize-expected.txt

This is a testharness.js-based test.
FAIL 'rotate(acos(0))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_not_equals: 'rotate(acos(0))' should be valid in transform. got disallowed value ""
FAIL 'rotate(acos(0))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'. assert_equals: 'rotate(acos(0))' and 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)' should serialize the same in computed values. expected "matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)" but got "none"
FAIL 'rotate(asin(1))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_not_equals: 'rotate(asin(1))' should be valid in transform. got disallowed value ""
FAIL 'rotate(asin(1))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'. assert_equals: 'rotate(asin(1))' and 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)' should serialize the same in computed values. expected "matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)" but got "none"
FAIL 'rotate(calc(acos(pi - pi)))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_not_equals: 'rotate(calc(acos(pi - pi)))' should be valid in transform. got disallowed value ""
FAIL 'rotate(calc(acos(pi - pi)))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'. assert_equals: 'rotate(calc(acos(pi - pi)))' and 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)' should serialize the same in computed values. expected "matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)" but got "none"
FAIL 'rotate(calc(asin(pi - pi + 1)))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_not_equals: 'rotate(calc(asin(pi - pi + 1)))' should be valid in transform. got disallowed value ""
FAIL 'rotate(calc(asin(pi - pi + 1)))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'. assert_equals: 'rotate(calc(asin(pi - pi + 1)))' and 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)' should serialize the same in computed values. expected "matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)" but got "none"
FAIL 'rotate(calc(atan(infinity)))' as a specified value should serialize as 'rotate(calc(90deg))'. assert_not_equals: 'rotate(calc(atan(infinity)))' should be valid in transform. got disallowed value ""
FAIL 'rotate(calc(atan(infinity)))' as a computed value should serialize as 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)'. assert_equals: 'rotate(calc(atan(infinity)))' and 'matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)' should serialize the same in computed values. expected "matrix(6.12323e-17, 1, -1, 6.12323e-17, 0, 0)" but got "none"
Harness: the test ran to completion.
